feat(nimbus): Move schemas
