+++ schemas/index.d.ts
@@ -7,7 +7,7 @@
  */
 
 /**
- * A unique, stable indentifier for the user used as an input to bucket hashing.
+ * A unique, stable identifier for the user used as an input to bucket hashing.
  */
 export type RandomizationUnit = "normandy_id" | "nimbus_id" | "user_id" | "group_id";
 export type DesktopApplication = "firefox-desktop" | "firefox-desktop-background-task";
@@ -49,114 +49,92 @@ export interface DesktopAllVersionsNimbusExperiment {
    */
   slug: string;
   /**
-   * Unique identifier for the experiiment.
-   *
-   * This is a duplicate of slug, but is required field for all Remote Settings records.
+   * Unique identifier for the experiment. This is a duplicate of slug.
    */
   id: string;
   /**
    * A slug identifying the targeted product of this experiment.
-   *
-   * It should be a lowercased_with_underscores name that is short and unambiguous and it should match the app_name found in https://probeinfo.telemetry.mozilla.org/glean/repositories. Examples are "fenix" and "firefox_desktop".
    */
   appName: string;
   /**
    * The platform identifier for the targeted app.
-   *
-   * This should match app's identifier exactly as it appears in the relevant app store listing (for relevant platforms) or the app's Glean initialization (for other platforms).
-   *
-   * Examples are "org.mozilla.firefox_beta" and "firefox-desktop".
    */
   appId: string;
   /**
-   * A specific channel of an application such as "nightly", "beta", or "release".
+   * A specific channel of an application such as 'nightly',             'beta', or 'release'.
    */
   channel: string;
   /**
-   * Public name of the experiment that will be displayed on "about:studies".
+   * Public name of the experiment displayed on 'about:studies'.
    */
   userFacingName: string;
   /**
-   * Short public description of the experiment that will be displayed on "about:studies".
+   * Short public description of the experiment.
    */
   userFacingDescription: string;
   /**
-   * When this property is set to true, the SDK should not enroll new users into the experiment that have not already been enrolled.
+   * Whether new users should be enrolled into the experiment.
    */
   isEnrollmentPaused: boolean;
   /**
-   * When this property is set to true, treat this experiment as a rollout.
-   *
-   * Rollouts are currently handled as single-branch experiments separated from the bucketing namespace for normal experiments.
-   *
-   * See-also: https://mozilla-hub.atlassian.net/browse/SDK-405
+   * Whether this experiment is a rollout.
    */
   isRollout?: boolean;
   bucketConfig: ExperimentBucketConfig;
   /**
-   * A list of outcomes relevant to the experiment analysis.
+   * List of outcomes relevant to analysis.
    */
   outcomes?: ExperimentOutcome[];
   /**
-   * A list of featureIds the experiment contains configurations for.
+   * List of featureIds the experiment includes.
    */
   featureIds?: string[];
   /**
-   * A JEXL targeting expression used to filter out experiments.
+   * A JEXL targeting expression.
    */
   targeting?: string | null;
   /**
    * Actual publish date of the experiment.
-   *
-   * Note that this value is expected to be null in Remote Settings.
    */
   startDate: string | null;
   /**
-   * Actual enrollment end date of the experiment.
-   *
-   * Note that this value is expected to be null in Remote Settings.
+   * Actual enrollment end date.
    */
   enrollmentEndDate?: string | null;
   /**
    * Actual end date of this experiment.
-   *
-   * Note that this field is expected to be null in Remote Settings.
    */
   endDate: string | null;
   /**
-   * Duration of the experiment from the start date in days.
-   *
-   * Note that this property is only used during the analysis phase (i.e., not by the SDK).
+   * Proposed duration of the experiment.
    */
   proposedDuration?: number;
   /**
-   * This represents the number of days that we expect to enroll new users.
-   *
-   * Note that this property is only used during the analysis phase (i.e., not by the SDK).
+   * Number of days expected for new user enrollment.
    */
   proposedEnrollment: number;
   /**
-   * The slug of the reference branch (i.e., the branch we consider "control").
+   * Slug of the reference branch, if applicable.
    */
   referenceBranch: string | null;
   /**
-   * The list of locale codes (e.g., "en-US" or "fr") that this experiment is targeting.
-   *
-   * If null, all locales are targeted.
+   * List of targeted locale codes.
    */
   locales?: string[] | null;
   /**
-   * The date that this experiment was first published to Remote Settings.
-   *
-   * If null, it has not yet been published.
+   * First published date to Remote Settings.
    */
   publishedDate?: string | null;
+  /**
+   * Per-locale localization substitutions.
+   */
+  localizations?: ExperimentLocalizations | null;
   /**
    * Branch configuration for the experiment.
    */
   branches: DesktopAllVersionsExperimentBranch[];
   /**
-   * When this property is set to true, treat this experiment as aFirefox Labs experiment
+   * Whether this experiment is a Firefox Labs Opt-In.
    */
   isFirefoxLabsOptIn?: boolean;
   /**
@@ -172,7 +150,7 @@ export interface DesktopAllVersionsNimbusExperiment {
    */
   firefoxLabsDescription?: string | null;
   /**
-   * Links that will be used with the firefoxLabsDescription Fluent ID. May be null for Firefox Labs Opt-In recipes that do not use links.
+   * Links that will be used with the Firefox Labs             description Fluent ID.
    */
   firefoxLabsDescriptionLinks?: {
     [k: string]: string;
@@ -183,12 +161,12 @@ export interface DesktopAllVersionsNimbusExperiment {
   featureValidationOptOut?: boolean;
   /**
    * Does the experiment require a restart to take effect?
-   *
-   * Only used by Firefox Labs Opt-Ins.
    */
   requiresRestart?: boolean;
-  localizations?: ExperimentLocalizations | null;
 }
+/**
+ * Common Bucketing Configuration used across all versions.
+ */
 export interface ExperimentBucketConfig {
   randomizationUnit: RandomizationUnit;
   /**
@@ -204,9 +182,7 @@ export interface ExperimentBucketConfig {
    */
   count: number;
   /**
-   * The total number of buckets.
-   *
-   * You can assume this will always be 10000
+   * The total number of buckets. You can assume             this will always be 10000.
    */
   total: number;
 }
@@ -221,6 +197,17 @@ export interface ExperimentOutcome {
   priority: string;
 }
 /**
+ * Per-locale localization substitutions.
+ *
+ * The top level key is the locale (e.g., "en-US" or "fr"). Each entry is a mapping of
+ * string IDs to their localized equivalents.
+ */
+export interface ExperimentLocalizations {
+  [k: string]: {
+    [k: string]: string;
+  };
+}
+/**
  * The branch definition supported on all Firefox Desktop versions.
  *
  * This version requires the feature field to be present to support older Firefox Desktop
@@ -233,8 +220,7 @@ export interface DesktopAllVersionsExperimentBranch {
   slug: string;
   /**
    * Relative ratio of population for the branch.
-   *
-   * e.g., if branch A=1 and branch B=3, then branch A would get 25% of the population.
+   * e.g., if branch A=1 and branch B=3, then branch A                 would get 25% of the population.
    */
   ratio: number;
   /**
@@ -254,8 +240,6 @@ export interface ExperimentFeatureConfig {
   featureId: string;
   /**
    * The values that define the feature configuration.
-   *
-   * This should be validated against a schema.
    */
   value: {
     [k: string]: unknown;
@@ -269,15 +253,275 @@ export interface DesktopPre95FeatureConfig {
   enabled: false;
 }
 /**
- * Per-locale localization substitutions.
+ * A Nimbus experiment for Firefox Desktop.
  *
- * The top level key is the locale (e.g., "en-US" or "fr"). Each entry is a mapping of
- * string IDs to their localized equivalents.
+ * This schema is less strict than DesktopAllVersionsNimbusExperiment and is intended for
+ * use in Firefox Desktop.
  */
-export interface ExperimentLocalizations {
-  [k: string]: {
+export interface DesktopNimbusExperiment {
+  /**
+   * Version of the NimbusExperiment schema this experiment refers to
+   */
+  schemaVersion: string;
+  /**
+   * Unique identifier for the experiment
+   */
+  slug: string;
+  /**
+   * Unique identifier for the experiment. This is a duplicate of slug.
+   */
+  id: string;
+  /**
+   * A slug identifying the targeted product of this experiment.
+   */
+  appName: string;
+  /**
+   * The platform identifier for the targeted app.
+   */
+  appId: string;
+  /**
+   * A specific channel of an application such as 'nightly',             'beta', or 'release'.
+   */
+  channel: string;
+  /**
+   * Public name of the experiment displayed on 'about:studies'.
+   */
+  userFacingName: string;
+  /**
+   * Short public description of the experiment.
+   */
+  userFacingDescription: string;
+  /**
+   * Whether new users should be enrolled into the experiment.
+   */
+  isEnrollmentPaused: boolean;
+  /**
+   * Whether this experiment is a rollout.
+   */
+  isRollout?: boolean;
+  bucketConfig: ExperimentBucketConfig;
+  /**
+   * List of outcomes relevant to analysis.
+   */
+  outcomes?: ExperimentOutcome[];
+  /**
+   * List of featureIds the experiment includes.
+   */
+  featureIds?: string[];
+  /**
+   * A JEXL targeting expression.
+   */
+  targeting?: string | null;
+  /**
+   * Actual publish date of the experiment.
+   */
+  startDate: string | null;
+  /**
+   * Actual enrollment end date.
+   */
+  enrollmentEndDate?: string | null;
+  /**
+   * Actual end date of this experiment.
+   */
+  endDate: string | null;
+  /**
+   * Proposed duration of the experiment.
+   */
+  proposedDuration?: number;
+  /**
+   * Number of days expected for new user enrollment.
+   */
+  proposedEnrollment: number;
+  /**
+   * Slug of the reference branch, if applicable.
+   */
+  referenceBranch: string | null;
+  /**
+   * List of targeted locale codes.
+   */
+  locales?: string[] | null;
+  /**
+   * First published date to Remote Settings.
+   */
+  publishedDate?: string | null;
+  /**
+   * Per-locale localization substitutions.
+   */
+  localizations?: ExperimentLocalizations | null;
+  /**
+   * Branch configuration for the experiment.
+   */
+  branches: DesktopExperimentBranch[];
+  /**
+   * Whether this experiment is a Firefox Labs Opt-In.
+   */
+  isFirefoxLabsOptIn?: boolean;
+  /**
+   * The group this should appear under in Firefox Labs
+   */
+  firefoxLabsGroup?: string | null;
+  /**
+   * The title shown in Firefox Labs (Fluent ID)
+   */
+  firefoxLabsTitle?: string | null;
+  /**
+   * The description shown in Firefox Labs (Fluent ID)
+   */
+  firefoxLabsDescription?: string | null;
+  /**
+   * Links that will be used with the Firefox Labs             description Fluent ID.
+   */
+  firefoxLabsDescriptionLinks?: {
     [k: string]: string;
-  };
+  } | null;
+  /**
+   * Opt out of feature schema validation.
+   */
+  featureValidationOptOut?: boolean;
+  /**
+   * Does the experiment require a restart to take effect?
+   */
+  requiresRestart?: boolean;
+}
+/**
+ * The branch definition supported on Firefox Desktop 95+.
+ */
+export interface DesktopExperimentBranch {
+  /**
+   * Identifier for the branch.
+   */
+  slug: string;
+  /**
+   * Relative ratio of population for the branch.
+   * e.g., if branch A=1 and branch B=3, then branch A                 would get 25% of the population.
+   */
+  ratio: number;
+  /**
+   * An array of feature configurations.
+   */
+  features: ExperimentFeatureConfig[];
+  /**
+   * The branch title shown in Firefox Labs (Fluent ID)
+   */
+  firefoxLabsTitle?: string | null;
+}
+/**
+ * A Nimbus experiment for Nimbus SDK-based applications.
+ */
+export interface SdkNimbusExperiment {
+  /**
+   * Version of the NimbusExperiment schema this experiment refers to
+   */
+  schemaVersion: string;
+  /**
+   * Unique identifier for the experiment
+   */
+  slug: string;
+  /**
+   * Unique identifier for the experiment. This is a duplicate of slug.
+   */
+  id: string;
+  /**
+   * A slug identifying the targeted product of this experiment.
+   */
+  appName: string;
+  /**
+   * The platform identifier for the targeted app.
+   */
+  appId: string;
+  /**
+   * A specific channel of an application such as 'nightly',             'beta', or 'release'.
+   */
+  channel: string;
+  /**
+   * Public name of the experiment displayed on 'about:studies'.
+   */
+  userFacingName: string;
+  /**
+   * Short public description of the experiment.
+   */
+  userFacingDescription: string;
+  /**
+   * Whether new users should be enrolled into the experiment.
+   */
+  isEnrollmentPaused: boolean;
+  /**
+   * Whether this experiment is a rollout.
+   */
+  isRollout?: boolean;
+  bucketConfig: ExperimentBucketConfig;
+  /**
+   * List of outcomes relevant to analysis.
+   */
+  outcomes?: ExperimentOutcome[];
+  /**
+   * List of featureIds the experiment includes.
+   */
+  featureIds?: string[];
+  /**
+   * A JEXL targeting expression.
+   */
+  targeting?: string | null;
+  /**
+   * Actual publish date of the experiment.
+   */
+  startDate: string | null;
+  /**
+   * Actual enrollment end date.
+   */
+  enrollmentEndDate?: string | null;
+  /**
+   * Actual end date of this experiment.
+   */
+  endDate: string | null;
+  /**
+   * Proposed duration of the experiment.
+   */
+  proposedDuration?: number;
+  /**
+   * Number of days expected for new user enrollment.
+   */
+  proposedEnrollment: number;
+  /**
+   * Slug of the reference branch, if applicable.
+   */
+  referenceBranch: string | null;
+  /**
+   * List of targeted locale codes.
+   */
+  locales?: string[] | null;
+  /**
+   * First published date to Remote Settings.
+   */
+  publishedDate?: string | null;
+  /**
+   * Per-locale localization substitutions.
+   */
+  localizations?: ExperimentLocalizations | null;
+  /**
+   * Branch configuration for the SDK experiment.
+   */
+  branches: SdkExperimentBranch[];
+}
+/**
+ * The branch definition for SDK-based applications.
+ *
+ * Supported on Firefox for Android 96+, Firefox for iOS 39+, and all versions of Cirrus.
+ */
+export interface SdkExperimentBranch {
+  /**
+   * Identifier for the branch.
+   */
+  slug: string;
+  /**
+   * Relative ratio of population for the branch.
+   * e.g., if branch A=1 and branch B=3, then branch A                 would get 25% of the population.
+   */
+  ratio: number;
+  /**
+   * An array of feature configurations.
+   */
+  features: ExperimentFeatureConfig[];
 }
 /**
  * A feature.
@@ -381,185 +625,6 @@ export interface DesktopFeatureManifest {
   [k: string]: DesktopFeature;
 }
 /**
- * A Nimbus experiment for Firefox Desktop.
- *
- * This schema is less strict than DesktopAllVersionsNimbusExperiment and is intended for
- * use in Firefox Desktop.
- */
-export interface DesktopNimbusExperiment {
-  /**
-   * Version of the NimbusExperiment schema this experiment refers to
-   */
-  schemaVersion: string;
-  /**
-   * Unique identifier for the experiment
-   */
-  slug: string;
-  /**
-   * Unique identifier for the experiiment.
-   *
-   * This is a duplicate of slug, but is required field for all Remote Settings records.
-   */
-  id: string;
-  /**
-   * A slug identifying the targeted product of this experiment.
-   *
-   * It should be a lowercased_with_underscores name that is short and unambiguous and it should match the app_name found in https://probeinfo.telemetry.mozilla.org/glean/repositories. Examples are "fenix" and "firefox_desktop".
-   */
-  appName: string;
-  /**
-   * The platform identifier for the targeted app.
-   *
-   * This should match app's identifier exactly as it appears in the relevant app store listing (for relevant platforms) or the app's Glean initialization (for other platforms).
-   *
-   * Examples are "org.mozilla.firefox_beta" and "firefox-desktop".
-   */
-  appId: string;
-  /**
-   * A specific channel of an application such as "nightly", "beta", or "release".
-   */
-  channel: string;
-  /**
-   * Public name of the experiment that will be displayed on "about:studies".
-   */
-  userFacingName: string;
-  /**
-   * Short public description of the experiment that will be displayed on "about:studies".
-   */
-  userFacingDescription: string;
-  /**
-   * When this property is set to true, the SDK should not enroll new users into the experiment that have not already been enrolled.
-   */
-  isEnrollmentPaused: boolean;
-  /**
-   * When this property is set to true, treat this experiment as a rollout.
-   *
-   * Rollouts are currently handled as single-branch experiments separated from the bucketing namespace for normal experiments.
-   *
-   * See-also: https://mozilla-hub.atlassian.net/browse/SDK-405
-   */
-  isRollout?: boolean;
-  bucketConfig: ExperimentBucketConfig;
-  /**
-   * A list of outcomes relevant to the experiment analysis.
-   */
-  outcomes?: ExperimentOutcome[];
-  /**
-   * A list of featureIds the experiment contains configurations for.
-   */
-  featureIds?: string[];
-  /**
-   * A JEXL targeting expression used to filter out experiments.
-   */
-  targeting?: string | null;
-  /**
-   * Actual publish date of the experiment.
-   *
-   * Note that this value is expected to be null in Remote Settings.
-   */
-  startDate: string | null;
-  /**
-   * Actual enrollment end date of the experiment.
-   *
-   * Note that this value is expected to be null in Remote Settings.
-   */
-  enrollmentEndDate?: string | null;
-  /**
-   * Actual end date of this experiment.
-   *
-   * Note that this field is expected to be null in Remote Settings.
-   */
-  endDate: string | null;
-  /**
-   * Duration of the experiment from the start date in days.
-   *
-   * Note that this property is only used during the analysis phase (i.e., not by the SDK).
-   */
-  proposedDuration?: number;
-  /**
-   * This represents the number of days that we expect to enroll new users.
-   *
-   * Note that this property is only used during the analysis phase (i.e., not by the SDK).
-   */
-  proposedEnrollment: number;
-  /**
-   * The slug of the reference branch (i.e., the branch we consider "control").
-   */
-  referenceBranch: string | null;
-  /**
-   * The list of locale codes (e.g., "en-US" or "fr") that this experiment is targeting.
-   *
-   * If null, all locales are targeted.
-   */
-  locales?: string[] | null;
-  /**
-   * The date that this experiment was first published to Remote Settings.
-   *
-   * If null, it has not yet been published.
-   */
-  publishedDate?: string | null;
-  /**
-   * Branch configuration for the experiment.
-   */
-  branches: DesktopExperimentBranch[];
-  /**
-   * When this property is set to true, treat this experiment as aFirefox Labs experiment
-   */
-  isFirefoxLabsOptIn?: boolean;
-  /**
-   * The group this should appear under in Firefox Labs
-   */
-  firefoxLabsGroup?: string | null;
-  /**
-   * The title shown in Firefox Labs (Fluent ID)
-   */
-  firefoxLabsTitle?: string | null;
-  /**
-   * The description shown in Firefox Labs (Fluent ID)
-   */
-  firefoxLabsDescription?: string | null;
-  /**
-   * Links that will be used with the firefoxLabsDescription Fluent ID. May be null for Firefox Labs Opt-In recipes that do not use links.
-   */
-  firefoxLabsDescriptionLinks?: {
-    [k: string]: string;
-  } | null;
-  /**
-   * Opt out of feature schema validation.
-   */
-  featureValidationOptOut?: boolean;
-  /**
-   * Does the experiment require a restart to take effect?
-   *
-   * Only used by Firefox Labs Opt-Ins.
-   */
-  requiresRestart?: boolean;
-  localizations?: ExperimentLocalizations | null;
-}
-/**
- * The branch definition supported on Firefox Desktop 95+.
- */
-export interface DesktopExperimentBranch {
-  /**
-   * Identifier for the branch.
-   */
-  slug: string;
-  /**
-   * Relative ratio of population for the branch.
-   *
-   * e.g., if branch A=1 and branch B=3, then branch A would get 25% of the population.
-   */
-  ratio: number;
-  /**
-   * An array of feature configurations.
-   */
-  features: ExperimentFeatureConfig[];
-  /**
-   * The branch title shown in Firefox Labs (Fluent ID)
-   */
-  firefoxLabsTitle?: string | null;
-}
-/**
  * The SDK-specific feature manifest.
  */
 export interface SdkFeatureManifest {
@@ -605,37 +670,6 @@ export interface SdkFeatureVariable {
    * Only allowed when type is string.
    */
   enum?: string[];
-}
-/**
- * A Nimbus experiment for Nimbus SDK-based applications.
- */
-export interface SdkNimbusExperiment {
-  /**
-   * Branch configuration for the experiment.
-   */
-  branches: SdkExperimentBranch[];
-}
-/**
- * The branch definition for SDK-based applications
- *
- * Supported on Firefox for Android 96+ and Firefox for iOS 39+ and all versions of
- * Cirrus.
- */
-export interface SdkExperimentBranch {
-  /**
-   * Identifier for the branch.
-   */
-  slug: string;
-  /**
-   * Relative ratio of population for the branch.
-   *
-   * e.g., if branch A=1 and branch B=3, then branch A would get 25% of the population.
-   */
-  ratio: number;
-  /**
-   * An array of feature configurations.
-   */
-  features: ExperimentFeatureConfig[];
 }
 export interface AnalysisError {
   analysis_basis?: AnalysisBasis | null;
