+++ schemas/mozilla_nimbus_schemas/__init__.py
@@ -1,2 +1,2 @@
-from mozilla_nimbus_schemas.experiments import *
+from mozilla_nimbus_schemas.experimenter_apis.experiments import *
 from mozilla_nimbus_schemas.jetstream import *
