+++ schemas/mozilla_nimbus_schemas/experimenter_apis/experiments/__init__.py
@@ -1,0 +1,28 @@
+from mozilla_nimbus_schemas.experimenter_apis.common import (
+    ExperimentBucketConfig,
+    ExperimentFeatureConfig,
+    ExperimentLocalizations,
+    ExperimentOutcome,
+    RandomizationUnit,
+)
+
+from .experiments import (
+    DesktopAllVersionsNimbusExperiment,
+    DesktopNimbusExperiment,
+    SdkNimbusExperiment,
+)
+from .feature_manifests import DesktopFeature, DesktopFeatureManifest, SdkFeatureManifest
+
+__all__ = (
+    "DesktopAllVersionsNimbusExperiment",
+    "DesktopNimbusExperiment",
+    "SdkNimbusExperiment",
+    "RandomizationUnit",
+    "ExperimentBucketConfig",
+    "ExperimentOutcome",
+    "ExperimentFeatureConfig",
+    "ExperimentLocalizations",
+    "DesktopFeature",
+    "DesktopFeatureManifest",
+    "SdkFeatureManifest",
+)
