+++ schemas/mozilla_nimbus_schemas/experimenter_apis/experiments_v7/__init__.py
@@ -1,12 +1,16 @@
+from mozilla_nimbus_schemas.experimenter_apis.common import (
+    ExperimentBucketConfig,
+    ExperimentFeatureConfig,
+    ExperimentLocalizations,
+    ExperimentOutcome,
+    RandomizationUnit,
+)
+
 from .experiments_v7 import (
     V7BaseExperiment,
     V7BaseExperimentBranch,
     V7DesktopExperimentBranch,
     V7DesktopNimbusExperiment,
-    V7ExperimentBucketConfig,
-    V7ExperimentFeatureConfig,
-    V7ExperimentOutcome,
-    V7RandomizationUnit,
     V7SdkExperimentBranch,
     V7SdkNimbusExperiment,
 )
@@ -15,11 +19,12 @@ __all__ = (
     "V7BaseExperiment",
     "V7DesktopNimbusExperiment",
     "V7SdkNimbusExperiment",
-    "V7RandomizationUnit",
-    "V7ExperimentBucketConfig",
-    "V7ExperimentOutcome",
-    "V7ExperimentFeatureConfig",
     "V7BaseExperimentBranch",
     "V7DesktopExperimentBranch",
     "V7SdkExperimentBranch",
+    "RandomizationUnit",
+    "ExperimentBucketConfig",
+    "ExperimentOutcome",
+    "ExperimentFeatureConfig",
+    "ExperimentLocalizations",
 )
