--- schemas/mozilla_nimbus_schemas/experiments/__init__.py
@@ -1,17 +1,0 @@
-from .experiments import (
-    DesktopAllVersionsNimbusExperiment,
-    DesktopNimbusExperiment,
-    RandomizationUnit,
-    SdkNimbusExperiment,
-)
-from .feature_manifests import DesktopFeature, DesktopFeatureManifest, SdkFeatureManifest
-
-__all__ = (
-    "DesktopAllVersionsNimbusExperiment",
-    "DesktopFeature",
-    "DesktopFeatureManifest",
-    "DesktopNimbusExperiment",
-    "RandomizationUnit",
-    "SdkFeatureManifest",
-    "SdkNimbusExperiment",
-)
