+++ schemas/schemas/V7SdkExperimentBranch.schema.json
@@ -15,7 +15,7 @@
     "features": {
       "description": "An array of feature configurations.",
       "items": {
-        "$ref": "#/$defs/V7ExperimentFeatureConfig"
+        "$ref": "#/$defs/ExperimentFeatureConfig"
       },
       "type": "array"
     }
@@ -26,7 +26,7 @@
     "features"
   ],
   "$defs": {
-    "V7ExperimentFeatureConfig": {
+    "ExperimentFeatureConfig": {
       "properties": {
         "featureId": {
           "description": "The identifier for the feature flag.",
